CloudSentinel/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
pydantic==2.7.4
pymongo==4.7.3
motor==3.4.0
pyyaml==6.0.1
python-dotenv==1.0.1
